% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a testing procedure}
\usage{
evaluate(data, ...)
}
\arguments{
\item{data}{A tibble, consists the \code{logical} columns \code{altr},
indicating whether the hypothesis is a true alternative, and
\code{final_result}, indicating whether the null hypothesis rejected
(considered as an alternative).}

\item{...}{\code{data}'s variables to group by.}
}
\value{
A tibble consists two columns, \code{frej} and \code{trej}, which are
the number of the false (true) rejections in each group.
}
\description{
Evaluate a testing procedure
}
