% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure}
\alias{measure}
\title{Measure a testing procedure by the FWER score}
\usage{
measure(data, ...)
}
\arguments{
\item{data}{A tibble, consists a \code{numeric} column, \code{frej}, which is
the number of the false rejections in each group.}

\item{...}{\code{data}'s variables to group by.}
}
\value{
A tibble, consists a \code{numeric} column, \code{fwer}, which is
the empirical probability of rejecting a true null.
}
\description{
Measure a testing procedure by the FWER score
}
