% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{perform}
\alias{perform}
\title{Perform a two-stage procedure}
\usage{
perform(data, filt_test, base_pval, base_thrl, base_thrl_adj_method, ...)
}
\arguments{
\item{data}{A tibble. Consists of raw observations, which may be
manipulated to construct filtration and base test statistics.}

\item{filt_test}{An expression, that its evaluation, with \code{data} as an
environment, produces a logical vector of length 1 or \code{data}'s number of
rows. A \code{TRUE} value correspond to an hypothesis that passes the filtration
test, i.e., is NOT filtered (filtration test rejected).}

\item{base_pval}{An expression, that its evaluation, with \code{data} as an
environment, produces a numeric vector of length 1 or \code{data}'s number of
rows. A \eqn{p}-value for the base test.}

\item{base_thrl}{An expression, that its evaluation, with \code{data} as an
environment, produces a numeric vector of length 1 or \code{data}'s number of
rows. A threshold, functioning as a reject region for the base test.}

\item{base_thrl_adj_method}{A function of 1 argument. Used to adjust the
thresholds of the base test, considering only unfiltered hypotheses.}

\item{...}{\code{data}'s variables to group by.}
}
\value{
\code{data}, with additional columns:
\itemize{
\item \code{filt_test}, \code{base_pval}, \code{base_thrl}: Evaluation
results of the corresponding parameter values.
\item \code{base_thrl_adj}: Adjusted thresholds for the base test; the
result of \code{base_thrl_adj_method} on the unfiltered thresholds in
\code{base_thrl}.
\item \code{final_result}: Logical, indicating for each hypothesis whether
both the filtration test and the base test of it are rejected (\code{TRUE}) or
not (\code{FALSE}).
}
}
\description{
Perform a two-stage procedure
}
