% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{reproduce}
\alias{reproduce}
\title{Reproduce the figures as they appear in the thesis.}
\usage{
reproduce(configid, output = c("trej", "fwer"))
}
\arguments{
\item{configid}{serial number of configuration, one of {1, 2, 3}.}

\item{output}{the desired plot, either "trej" for number of true rejections,
or "fwer" for the FWER.}
}
\value{
the plot of the \code{output}, under the \code{configid}-th
configuration
}
\description{
Reproduce the figures as they appear in the thesis.
}
\examples{
reproduce(1, "trej")
reproduce(2, "fwer")

}
