% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_estimate.R
\name{shrink_estimate}
\alias{shrink_estimate}
\title{Estimate parameter with shrinkage estimator}
\usage{
shrink_estimate(data, estim_base, estim_filt, filtfactor, filtexp)
}
\arguments{
\item{data}{A tibble. Consists of raw observations, which may be
manipulated to construct filtration and base test statistics.}

\item{estim_base}{An expression, that its evaluation, with \code{data} as an
environment, produces a numeric vector of length 1 or \code{data}'s number of
rows. A base statistic.}

\item{estim_filt}{An expression, that its evaluation, with \code{data} as an
environment, produces a numeric vector of length 1 or \code{data}'s number of
rows. A filtration statistic.}

\item{filtfactor}{A \code{numeric} constant, used to define the filtration
threshold.}

\item{filtexp}{A \code{numeric} constant, used to define the filtration
threshold.}
}
\value{
\code{data}, with additional columns:
\itemize{
\item \code{estim_base}, \code{estim_filt}: Evaluation
results of the corresponding parameter values.
\item \code{filt}: Whether or not a filtration has occurred:
\code{estim_filt < filtfactor * nobs^(-filtexp)}.
\item \code{estim_shrink}: Shrinkage estimator's value.
}
}
\description{
Estimate parameter with shrinkage estimator
}
