% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_evaluate.R
\name{shrink_evaluate}
\alias{shrink_evaluate}
\title{Evaluate a shrinkage estimator}
\usage{
shrink_evaluate(data, ...)
}
\arguments{
\item{data}{A tibble, consists the \code{numeric} columns \code{estim_base}
and \code{estim_shrink}, the two estimators, and \code{param}, the real
parameter.}
}
\value{
A tibble consists of MSE of the two estimators, and the MSE-ratio of
them.
}
\description{
Evaluate a shrinkage estimator
}
