% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_reproduce.R
\name{shrink_reproduce}
\alias{shrink_reproduce}
\title{Reproduce figures}
\usage{
shrink_reproduce(configid)
}
\arguments{
\item{configid}{Configuration identifier.}
}
\value{
Plot of the MSE-ratio.
}
\description{
Reproduce figures
}
\examples{
shrink_reproduce(1)
}
