% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_simulate.R
\name{shrink_simulate}
\alias{shrink_simulate}
\title{Simulate data for mediation estimation}
\usage{
shrink_simulate(nobs, nrep, config)
}
\arguments{
\item{nobs}{Number of observations (may be a vector)}

\item{nrep}{Number of experiments for each \code{nobs} value.}

\item{config}{A data frame, consists information about the different cases
(values of true parameters) in the following columns: \code{gfactor} and
\code{bfactor}; and \code{gexp} and \code{bexp}: \code{numeric}
constants used to define the true parameters \eqn{\gamma_n} and
\eqn{\beta_n}.}
}
\value{
A tibble with \code{length(nobs) * nrep * nrows(config)} rows. Each row
corresponds to an realization of an estimator, assumed to be an
approximation to an asymptotically normal estimator with root-\eqn{n} rate
(based on \eqn{n} observations). The simulations here are based on the
normal approximations, and consists of the following columns:
\itemize{
\item \code{nobs}: Number of observations (affecting the standard
deviation). Values are from the given value in the corresponding
parameter \code{nobs}.
\item \code{case}: Identifier of the true case of the hypothesis.
\item \code{g} and \code{b}: The true parameters. Their values depends
on the parameters in the config, and on the current \code{nobs} value:
\eqn{\gamma_n} = \code{gfactor * nobs^gexp}; \eqn{\beta_n} =
\code{bfactor * nobs^bexp}.
\item \code{gestim} and \code{bestim}: estimators for \eqn{\gamma_n} and
\eqn{\beta_n}, distributed independently from Normal distribution with
mean \eqn{\gamma_n} and \eqn{\beta_n} respectively, and with standard
deviation equals \eqn{1/\sqrt n}.
}
\description{
Simulate data for mediation estimation
}
