% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate data for mediation hypotheses}
\usage{
simulate(nobs, nexper, nhyp, config)
}
\arguments{
\item{nobs}{Number of observations (may be a vector)}

\item{nexper}{Number of experiments for each \code{nobs} value.}

\item{nhyp}{Total number of hypotheses (nulls and alternatives) in each
experiment.}

\item{config}{A data frame, consists information about the different cases
(values of true parameters) in the following columns:
\itemize{
\item \code{altr}: \code{logical}, indicating whether this is a case of
the alternative.
\item \code{gaddend} and \code{baddend}; \code{gfactor} and
\code{bfactor}; and \code{gexp} and \code{bexp}: \code{numeric}
constants used to define the true parameters \eqn{\gamma_n} and
\eqn{\beta_n}.
\item \code{prop}: \code{numeric}, discrete distribution for the
different cases in the configuration.
}}
}
\value{
A tibble with \code{length(nobs) * nexper * nhyp} rows. Each row
corresponds to an realization of an estimator, assumed to be an
approximation to an asymptotically normal estimator with root-\eqn{n} rate
(based on \eqn{n} observations). The simulations here are based on the
normal approximations, and consists of the following columns:
\itemize{
\item \code{nobs}: Number of observations (affecting the standard
deviation). Values are from the given value in the corresponding
parameter \code{nobs}.
\item \code{exper}: Serial number of the experiment in the current
\code{nobs} value.
\item \code{case}: Identifier of the true case of the hypothesis. In each
experiment the hypotheses are distributed according to \code{config}.
\item \code{g} and \code{b}: The true parameters. Their values depends
on the parameters in the config, and on the current \code{nobs} value:
\eqn{\gamma_n} = \code{gaddend + gfactor * nobs^gexp}; \eqn{\beta_n} =
\code{baddend + bfactor * nobs^bexp}.
\item \code{gestim} and \code{bestim}: estimators for \eqn{\gamma_n} and
\eqn{\beta_n}, distributed independently from Normal distribution with
mean \eqn{\gamma_n} and \eqn{\beta_n} respectively, and with standard
deviation equals \eqn{1/\sqrt n}.
\item \code{gpval} and \code{bpval}: Corresponding \eqn{p}-values,
under the null hypotheses \eqn{\gamma = 0} and \eqn{\beta = 0},
respectively.
}
\description{
Simulate data for mediation hypotheses
}
